Extract category view renderers
